Emulate writing of toy tags
Added a hook to the toypad that writes data to the tag in the top position when prompted. The data is saved, but in the current state this data is immediately forgotten if the tag is moved. However it does allow you to pass forced vehicle builds in the story and levels.